Display score breakdown on end game screen
Fixes #271
